Removed home page
It doesn't seem to serve much of a purpose in it's current state.
