--- app/homes/page.module.css
@@ -1,33 +1,0 @@
-@value smallAndUp from '../../styles/breakpoints.module.css';
-
-.search {
-  /* Requires all parent elements to also be height: 100% in order to work */
-  height: 100%;
-  display: grid;
-  grid-template-rows: auto 1fr;
-}
-
-.results {
-  height: 100%;
-  position: relative;
-}
-
-@media smallAndUp {
-  .results {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    /* Setting min-height/min-width prevents the results container from
-    overflowing off the screen when the content of .searchResults overflows the
-    window height. We only want .searchResults to scroll, not the entire window
-    */
-    min-height: 0;
-    min-width: 0;
-    padding-bottom: 1rem;
-  }
-}
-
-@media largeAndUp {
-  .results {
-    grid-template-columns: 55% 1fr;
-  }
-}
