+++ app/page.tsx
@@ -1,58 +1,33 @@
-import Link from 'next/link'
-import HomePageHeader from '../components/header/HomePageHeader/HomePageHeader'
-import LoginOrRegisterModal from '../containers/modals/LoginOrRegisterModal/LoginOrRegisterModal'
-import StandaloneSearchField from '../containers/StandaloneSearchField/StandaloneSearchField'
+'use client'
+
+import type { NextPage } from 'next'
+import { useGetCurrentUserIfAuthenticated } from '../hooks/get_current_user_if_authenticated_hook'
+import { useSyncSearchStateWithUrl } from '../hooks/sync_search_state_with_url_hook'
 import GoogleMapsProvider from '../providers/GoogleMapsProvider'
+import SearchHeader from '../containers/SearchHeader/SearchHeader'
+import SearchResults from '../containers/SearchResults/SearchResults'
+import ListingMap from '../containers/ListingMap/ListingMap'
+import SearchModals from '../components/SearchModals'
 import styles from './page.module.css'
 
-const HomePage: React.FC = () => {
+const SearchPage: NextPage = () => {
+  // We want to get the currentUser so that we can display their favorites on
+  // the listing cards
+  useGetCurrentUserIfAuthenticated()
+  useSyncSearchStateWithUrl()
+
   return (
-    <GoogleMapsProvider libraries={['places']}>
-      <div className={styles.home}>
-        <div>
-          <HomePageHeader />
+    <GoogleMapsProvider>
+      <div className={styles.search}>
+        <SearchHeader />
+        <div className={styles.results}>
+          <SearchResults />
+          <ListingMap />
         </div>
-
-        <div className={styles.body}>
-          <div className={styles.hero}>
-            <h1>Discover the Awsöm!</h1>
-            <StandaloneSearchField />
-          </div>
-
-          <main className={styles.mainContent}>
-            <h2>Explore homes on Awsom</h2>
-            <p>
-              Check out awsöm homes with our patented{' '}
-              <em>&quot;Search That Doesn&apos;t Suck™</em>&quot;. A brand new
-              search that looks cool and skips all the features you don&apos;t
-              really use.
-            </p>
-
-            <h2>Hot Searches in Seattle 🔥</h2>
-            <ul>
-              <li>
-                Checkout{' '}
-                <Link href='/homes?location=Fremont--Seattle--WA--USA'>
-                  Fremont
-                </Link>
-                , a great neighborhood that you probably can&apos;t afford, but
-                it does&apos;t hurt to look!
-              </li>
-              <li>
-                Search for{' '}
-                <Link href='/homes?location=Ballard--Seattle--WA--USA&beds-min=2&baths-min=1'>
-                  {' '}
-                  Ballard 2+bd 1+ba
-                </Link>{' '}
-                but be prepared to pay a lot of 💰💰💰💰💰💰💰💰 :(
-              </li>
-            </ul>
-          </main>
-        </div>
-        <LoginOrRegisterModal />
+        <SearchModals />
       </div>
     </GoogleMapsProvider>
   )
 }
 
-export default HomePage
+export default SearchPage
